move stabilization to interface
