+++ include/aspect/mesh_deformation/interface.h
@@ -24,6 +24,7 @@
 
 #include <aspect/plugins.h>
 #include <aspect/simulator_access.h>
+#include <aspect/simulator/assemblers/interface.h>
 
 #include <aspect/global.h>
 
@@ -41,6 +42,36 @@ namespace aspect
   using namespace dealii;
 
   template <int dim> class Simulator;
+
+namespace Assemblers
+  {
+    /**
+     * Apply stabilization to a cell of the system matrix. The
+     * stabilization is only added to cells on a free surface. The
+     * scheme is based on that of Kaus et. al., 2010. Called during
+     * assembly of the system matrix.
+     */
+    template <int dim>
+    class ApplyStabilization: public Assemblers::Interface<dim>,
+      public SimulatorAccess<dim>
+    {
+      public:
+        ApplyStabilization(const double stabilization_theta);
+
+        void
+        execute (internal::Assembly::Scratch::ScratchBase<dim>   &scratch,
+                 internal::Assembly::CopyData::CopyDataBase<dim> &data) const override;
+
+      private:
+        /**
+         * Stabilization parameter for the free surface. Should be between
+         * zero and one. A value of zero means no stabilization. See Kaus
+         * et. al. 2010 for more details.
+         */
+        const double surface_theta;
+    };
+  }
+
 
   /**
    * A namespace that contains everything that is related to the deformation
@@ -160,6 +191,13 @@ namespace aspect
          * The default implementation of this function does nothing.
          */
         void initialize();
+
+      /**
+       * Called by Simulator::set_assemblers() to allow the FreeSurface plugin
+       * to register its assembler.
+       */
+      void set_assemblers(const SimulatorAccess<dim> &simulator_access,
+                          aspect::Assemblers::Manager<dim> &assemblers) const;
 
         /**
          * Update function of the MeshDeformationHandler. This function
@@ -333,6 +371,8 @@ namespace aspect
          * surface are fixed according to the selected deformation plugins.
          */
         AffineConstraints<double> make_initial_constraints ();
+
+        double surface_theta;
 
         /**
          * Deform the initial mesh by solving a Laplace equation
